Tests, reorganize models
