--- src/app/models/belt.ts
@@ -1,13 +1,0 @@
-import { Rational } from './rational';
-
-export interface Belt {
-  speed: number;
-}
-
-export class RationalBelt {
-  speed: Rational;
-
-  constructor(data: Belt) {
-    this.speed = Rational.fromNumber(data.speed);
-  }
-}
