+++ src/app/models/data/belt.ts
@@ -1,0 +1,13 @@
+import { Rational } from '../rational';
+
+export interface Belt {
+  speed: number;
+}
+
+export class RationalBelt {
+  speed: Rational;
+
+  constructor(data: Belt) {
+    this.speed = Rational.fromNumber(data.speed);
+  }
+}
