+++ src/app/models/enum/display-rate.ts
@@ -1,0 +1,13 @@
+import { Rational } from '../rational';
+
+export enum DisplayRate {
+  PerSecond = 1,
+  PerMinute = 60,
+  PerHour = 3600,
+}
+
+export const DisplayRateVal = {
+  [DisplayRate.PerSecond]: new Rational(BigInt(DisplayRate.PerSecond)),
+  [DisplayRate.PerMinute]: new Rational(BigInt(DisplayRate.PerMinute)),
+  [DisplayRate.PerHour]: new Rational(BigInt(DisplayRate.PerHour)),
+};
