+++ src/app/models/enum/rate-type.ts
