--- src/app/models/recipe.ts
@@ -1,80 +1,0 @@
-import { ItemId } from './item';
-import { Rational } from './rational';
-import { Entities } from './entities';
-
-export enum RecipeId {
-  AdvancedOilProcessing = 'advanced-oil-processing',
-  BasicOilProcessing = 'basic-oil-processing',
-  CoalLiquefaction = 'coal-liquefaction',
-  ElectronicCircuit = 'electronic-circuit',
-  HeavyOilCracking = 'heavy-oil-cracking',
-  KovarexEnrichmentProcess = 'kovarex-enrichment-process',
-  LightOilCracking = 'light-oil-cracking',
-  RocketPart = 'rocket-part',
-  Satellite = 'satellite',
-  SolidFuelFromLightOil = 'solid-fuel-from-light-oil',
-  SolidFuelFromPetroleumGas = 'solid-fuel-from-petroleum-gas',
-  SpaceSciencePack = 'space-science-pack',
-  SteelChest = 'steel-chest',
-  UraniumProcessing = 'uranium-processing',
-}
-
-export interface Recipe {
-  id: RecipeId;
-  name: string;
-  time: number;
-  in?: Entities<number>;
-  out?: Entities<number>;
-  expensive?: {
-    time?: number;
-    in?: Entities<number>;
-  };
-  producers?: ItemId[];
-}
-
-export class RationalRecipe {
-  id: RecipeId;
-  name: string;
-  time: Rational;
-  in?: Entities<Rational>;
-  out?: Entities<Rational>;
-  expensive?: {
-    time?: Rational;
-    in?: Entities<Rational>;
-  };
-  producers?: ItemId[];
-
-  constructor(data: Recipe) {
-    this.id = data.id;
-    this.name = data.name;
-    this.time = Rational.fromNumber(data.time);
-    if (data.in) {
-      this.in = Object.keys(data.in).reduce((e: Entities<Rational>, i) => {
-        return { ...e, ...{ [i]: Rational.fromNumber(data.in[i]) } };
-      }, {});
-    }
-    if (data.out) {
-      this.out = Object.keys(data.out).reduce((e: Entities<Rational>, i) => {
-        return { ...e, ...{ [i]: Rational.fromNumber(data.out[i]) } };
-      }, {});
-    }
-    if (data.expensive) {
-      this.expensive = {};
-      if (data.expensive.time) {
-        this.expensive.time = Rational.fromNumber(data.expensive.time);
-      }
-      if (data.expensive.in) {
-        this.expensive.in = Object.keys(data.expensive.in).reduce(
-          (e: Entities<Rational>, i) => {
-            return {
-              ...e,
-              ...{ [i]: Rational.fromNumber(data.expensive.in[i]) },
-            };
-          },
-          {}
-        );
-      }
-    }
-    this.producers = data.producers;
-  }
-}
